Remove unnecessary req.session.save calls
clearRouteData should always be at the end of the
middleware chain and upon res end the session will
save automatically.

The other routes also called req.session.save,
but the changes to the session should persist
from middleware to middleware for the duration of
the response. All the calls to req.session.save
just make for numerous unnecessary database writes.